+++ css/style.css
@@ -1437,6 +1437,6 @@ main {
   background-color: #ff0000;
 }
 
-.temp {
+.photos_container {
   display: contents;
 }
