+++ index.html
@@ -133,7 +133,7 @@
       </section>
 
       <!-- Здесь будут изображения других пользователей -->
-      <section class="pictures  container  temp">
+      <section class="pictures container photos_container">
 
       </section>
     </section>
